+++ scripts/compose/main.go
@@ -265,7 +265,7 @@ func generateQ3() {
 		fmt.Printf("  q3-group-%v:\n", i)
 		fmt.Printf("    container_name: q3-group-%v\n", i)
 		fmt.Println("    image: tp1:latest")
-		fmt.Println("    entrypoint: /group-by-game")
+		fmt.Println("    entrypoint: /build/group-by")
 		fmt.Println("    environment:")
 		fmt.Println("      - RABBIT_IP=rabbitmq")
 		fmt.Printf("      - PARTITION_ID=%v\n", i)
@@ -337,7 +337,7 @@ func generateQ4() {
 		fmt.Printf("  q4-group-%v:\n", i)
 		fmt.Printf("    container_name: q4-group-%v\n", i)
 		fmt.Println("    image: tp1:latest")
-		fmt.Println("    entrypoint: /group-by-game")
+		fmt.Println("    entrypoint: /build/group-by")
 		fmt.Println("    environment:")
 		fmt.Println("      - RABBIT_IP=rabbitmq")
 		fmt.Printf("      - PARTITION_ID=%v\n", i)
@@ -410,7 +410,7 @@ func generateQ5() {
 		fmt.Printf("  q5-group-%v:\n", i)
 		fmt.Printf("    container_name: q5-group-%v\n", i)
 		fmt.Println("    image: tp1:latest")
-		fmt.Println("    entrypoint: /group-by-game")
+		fmt.Println("    entrypoint: /build/group-by")
 		fmt.Println("    environment:")
 		fmt.Println("      - RABBIT_IP=rabbitmq")
 		fmt.Printf("      - PARTITION_ID=%v\n", i)
